+++ trex/src/aggregators.rs
@@ -1,4 +1,4 @@
-use std::f32;
+use std::f64;
 use std::ops::Add;
 use std::ops::Deref;
 use tesla::{AttributeDeclaration, Event};
@@ -15,13 +15,13 @@ fn compute_average<'a, T, U>(iterator: T,
     match attributes[attr].ty {
         BasicType::Int => {
             let mapped = iterator.map(|evt| evt.tuple.data[attr].unwrap_int());
-            let (count, sum) = mapped.fold((0i32, 0), |acc, x| (acc.0 + 1, acc.1 + x));
-            if count > 0 { Some(Value::from(sum as f32 / count as f32)) } else { None }
+            let (count, sum) = mapped.fold((0i64, 0), |acc, x| (acc.0 + 1, acc.1 + x));
+            if count > 0 { Some(Value::from(sum as f64 / count as f64)) } else { None }
         }
         BasicType::Float => {
             let mapped = iterator.map(|evt| evt.tuple.data[attr].unwrap_float());
-            let (count, sum) = mapped.fold((0i32, 0.0), |acc, x| (acc.0 + 1, acc.1 + x));
-            if count > 0 { Some(Value::from(sum / count as f32)) } else { None }
+            let (count, sum) = mapped.fold((0i64, 0.0), |acc, x| (acc.0 + 1, acc.1 + x));
+            if count > 0 { Some(Value::from(sum / count as f64)) } else { None }
         }
         _ => panic!("Tring to compute aggregate on wrong Value type"),
     }
@@ -61,7 +61,7 @@ fn compute_min<'a, T, U>(iterator: T,
         }
         BasicType::Float => {
             let mapped = iterator.map(|evt| evt.tuple.data[attr].unwrap_float());
-            let min = mapped.fold(f32::NAN, f32::min);
+            let min = mapped.fold(f64::NAN, f64::min);
             if !min.is_nan() { Some(Value::from(min)) } else { None }
         }
         _ => panic!("Tring to compute aggregate on wrong Value type"),
@@ -82,7 +82,7 @@ fn compute_max<'a, T, U>(iterator: T,
         }
         BasicType::Float => {
             let mapped = iterator.map(|evt| evt.tuple.data[attr].unwrap_float());
-            let max = mapped.fold(f32::NAN, f32::max);
+            let max = mapped.fold(f64::NAN, f64::max);
             if !max.is_nan() { Some(Value::from(max)) } else { None }
         }
         _ => panic!("Tring to compute aggregate on wrong Value type"),
@@ -101,6 +101,6 @@ pub fn compute_aggregate<'a, T, U>(aggregator: &Aggregator,
         Aggregator::Sum(attr) => compute_sum(iterator, attributes, attr),
         Aggregator::Min(attr) => compute_min(iterator, attributes, attr),
         Aggregator::Max(attr) => compute_max(iterator, attributes, attr),
-        Aggregator::Count => Some(Value::from(iterator.count() as i32)),
+        Aggregator::Count => Some(Value::from(iterator.count() as i64)),
     }
 }
